build: fix build with -Wl,--as-needed
Fixes: oneapi-src/oneVPL-intel-gpu#111

-Wl,--no-as-needed is actually a GNU Linker default (see man 1 ld), we
don't need to explicitly specify it on a linker cmdline. Especially
considering that some Linux distros on purpose build with the
-Wl,--as-needed instead.

-Wl,--as-needed looks to be stronger requirement than -Wl,--no-as-needed,
so it seems reasonable to use this as a default for our CI.

Finally, below build issue which we encounter with -Wl,--as-needed is
caused by missing -ldl for the internal library where we
actually use dlopen() stuff. We did have -ldl on a cmdline, but it was
at a wrong location, i.e. was added by cmake earlier than actually needed,
which caused link error.

[100%] Linking CXX shared library ../../__bin/None/libmfx-gen.so
/usr/bin/ld: ../../__lib/None/libvm.a(vm_shared_object_linux32.c.o): in function `vm_so_load':
vm_shared_object_linux32.c:(.text+0xc): undefined reference to `dlopen'

Signed-off-by: Dmitry Rogozhkin <dmitry.v.rogozhkin@intel.com>