[VPL] Fix Pre-Si platform detection failure issue (#3923)
Currently there is failure when detect some derivative platforms which reuses the same HW type of leading platform. This patch uses both platform HW type and Device Id to detect if it is pre-si environment. Need to delete the Device Id judgement codes when one platform changed to be post-si paltform.

Co-authored-by: Hao Chen <hao5.chen@intel.com>