[VPL] Fix extBrc and MCTF don't work on ADL platform issue (#4565)
Root cause is ADL platforms are missed in ASC CM and MCTF support list. ASC CM is used in extBrc and MCTF is a subfeature of VME. Both ASC CM and VME are only supported on platforms before DG2. This patch is to enable those two features on ADL platform.

Co-authored-by: Hao Chen <hao5.chen@intel.com>